Migrate to Manifest V3
