+++ src/content/loader.ts
@@ -1,0 +1,5 @@
+import Loader from "../components/loader.svelte";
+
+new Loader({
+    target: document.getElementById("ptt-loader"),
+});
