+++ src/content/pages/Page.ts
@@ -1,0 +1,3 @@
+export interface Page {
+    initialise(): Promise<void>;
+}
