+++ src/content/utils.ts
@@ -1,7 +1,7 @@
 export const buildLink = (arg: string) =>
     new URL(
         "index.php?" + arg,
-        window.location.protocol + "//" + window.location.hostname,
+        window.location.protocol + "//" + window.location.host,
     );
 
 export function getCourseId() {
