+++ src/events.ts
@@ -1,0 +1,1 @@
+export const pttLoadedEvent = new Event("pttLoaded");
