+++ src/messages.ts
@@ -1,3 +1,4 @@
 export enum MessageType {
     GET_SETTINGS = "GET_SETTINGS",
+    GET_LOADER = "GET_LOADER",
 }
