+++ src/types.d.ts
@@ -1,8 +1,4 @@
 interface Window {
-    progtestThemes?: {
-        loaded?: boolean;
-    };
-
     // TODO: Migrate to using highlight.js when each page is in its own file
     hljs?: HLJSApi;
 }
